Add experimental conditional syntax to text compiler
This is a first attempt, just to get *something* working. The parsing is
messy and unfinished, but should provide enough of a starting point to
explore different variations of template syntax and decide which one
works best in the context of prose text. This feature isn’t currently
being used in Muturangi, so it’s a bit more speculative for now (though
there are a few places where I really want to use this capability).