Use tilde to demarcate directives instead of asterisks
Yes. Changed my mind again, after playing with the syntax some more. I
don’t want to introduce scope wrappers if I can avoid it, but asterisks
are generally used for italics and bolds in Markdown and other formats,
so it’s probably a bad (or confusing) idea to change their meaning here.